IMPROVED: Major update

+++ ir_attachment_orphan_gc/models/ir_attachment.py
@@ -2,15 +2,15 @@
 
 import logging
 from odoo import fields, models
-from odoo.tools.profiler import profile
+# from odoo.tools.profiler import profile
 
 _logger = logging.getLogger(__name__)
 
 class IrAttachment(models.Model):
     _inherit = 'ir.attachment'
 
-    CG_MODULE_NAME_BLACKLIST = []
-    CG_MODULE_PREFIX_BLACKLIST = ['ir.']
+    CG_MODULE_NAME_SAFELIST = []
+    CG_MODULE_PREFIX_SAFELIST = ['ir.']
 
     maybe_orphan = fields.Boolean(
         readonly = True,
@@ -21,71 +21,24 @@ class IrAttachment(models.Model):
             will be removed.'''
     )
 
-    # @todo: ?? Aggiungere constraint che dà errore se si marcano record con "res_id" != False ??
-
-    # @todo: Possibilità di eseguirlo manualmente solo su un modello senza che questo sia attivo,
-    #        passando il nome del modello come argomento. Ovviamente bisogna eseguirlo due volte per
-    #        eliminare definitivemente gli attachemnt. Eseguendolo maualmente però i file fisici
-    #        saranno eliminati al prossimo vacuum cron (power_on).
-    #        Magari aggiungere il pulsante all'action nel form view di "ir.model".
-    #        Se eseguiita manualmente, mostra info ad-hoc nel log.
-
-    # @todo: Implementare registro/log degli attachment cancellati (hash, tipo, dimensione, res_name, res_model)
-
-    # @todo: Implementare l'esecuzione solo un giorno alla settimana (es. domenica)
-
-    # @todo: Implementare action multi "set as not orphan"
-
-    '''
-       Nei modelli in cui la funzionalità è attiva, elimina tutti gli attachment rimasti orfani
-       dopo che tutte le relazioni (con qualunque altro modello nell'environment) sono
-       state interrotte a causa dell'eliminazione dei record oppure per la modifica dei campi
-       che puntavano all'attachment. Questo problema si presenta in particolar modo utilizzando
-       il widget "many2many_binary".
-       
-       Questo garbage collector elimina solo gli attachment con i campi "res_id" vuoto e "res_model"
-       che punta ad un modello il cui attributo "_attachment_garbage_collector" impostato a True.
-
-       ATTENZIONE:
-       Gestisce solo campi Many2any e Many2one che puntano a record in "ir.attachment"
-       e con dominio [('res_model', '=', _name)]. In altre parole, se attivi la
-       funzione su un modulo e vuoi che elimini TUTTI gli attachment rimesti orfani,
-       non dovresti puntare ad attachment il cui "res_model" non sia anche esso un modulo
-       con  "_attachment_garbage_collector" impostato a True.
-       
-       Per i campi Many2many si presume di usare il widget "many2many_binary"
-       il quale si occupa di gestire correttamente il dominio degli attachment
-       visualizzati e i valori di default di quelli creati (soprattutto "res_model").
-
-       Con i campi Many2one si può aggiungere eventualmente la condizione
-       [('res_id', '=', id)] al dominio precedente (il quale resta imprescindibile);
-       in questo caso è necessario ereditare il 'mail.thread' mixin il quale si occuperà
-       lui di gestire/eliminare nel modo corretto gli attachment con [('res_id', '!=', 0)].
-
-       Alcuni modelli hanno una gestione degli attachment non standard e potrebbero
-       verificarsi effetti inattesi se questo garbage collector venisse attivato su
-       di essi. In questo caso basta aggiungere il nome dei modelli da ignorare
-       al dict CG_MODULE_NAME_BLACKLIST per evitare un'attivazione erronea su di essi.
-       Es. self.env['ir.attachment'].CG_MODULE_NAME_BLACKLIST.append('model.to.add')
-
-       La prima volta che individua degli attachment orfani li marchia solo (campo
-       "maybe_orphan"). La seconda volta che questo metodo viene eseguito, se trova che
-       sono gli stessi, li elimina (questo per ridurre possibili errori quando il numero
-       di attachment sarà molto grande).
-       Quelli marchiati precedentemente e che non risultano più orfani, vengono de-marchiati.
-
-    '''
-
     # @profile
-    def _garbage_collect_rel_orphan_attachments(self):
+    def _garbage_collect_rel_orphan_attachments(self, do_unlink=False, force_models=False):
+        """ Performs the check of attachments in "ir.attachment" according to the parameters passed.
+
+            :param bool do_unlink: If True perform unlink(), if False mark only.
+            :param list force_models: List of strings containing the names of the models to be checked.
+            :return: True
+        """
+
+        _logger.info('ATTACHMENTS GC - STARTED' + ('.' if do_unlink else ' in safe mode.'))
         count = {'model': 0, 'm2m': 0, 'm2o': 0}
 
-        # Prepara la blacklist effettiva
-        cg_name_blacklist = [ model_name
-            for model_name in self.env if model_name[0:3] in self.CG_MODULE_PREFIX_BLACKLIST
+        # Prepara la safelist effettiva
+        cg_name_safelist = [ model_name
+            for model_name in self.env if model_name[0:3] in self.CG_MODULE_PREFIX_SAFELIST
         ]
-        cg_name_blacklist = list(set(cg_name_blacklist + self.CG_MODULE_NAME_BLACKLIST))
-        _logger.info('ATTACHMENTS GARBAGE COLLECTOR - The model blackist is: %s' % ', '.join(cg_name_blacklist))
+        cg_name_safelist = list(set(cg_name_safelist + self.CG_MODULE_NAME_SAFELIST))
+        _logger.info('ATTACHMENTS GC - The model safelist is: %s' % ', '.join(cg_name_safelist))
 
         # Tutti i modelli che hanno campi Many2many o Many2one che puntano a "ir.attachemnt"
         rel_models = []
@@ -121,16 +74,20 @@ class IrAttachment(models.Model):
             else:
                 continue
 
-            # Legge l'attributo che attiva il gc sul modello
-            is_active_model = getattr(Model, '_attachment_garbage_collector', False)
-            # Valida il tipo dell'attributo
-            if type(is_active_model) is not bool:
+            if force_models:
+                is_active_model = True if model_name in force_models else False
+            else: 
+                # Legge l'attributo che attiva il gc sul modello
+                is_active_model = getattr(Model, '_attachment_garbage_collector', False)
+                # Valida il tipo dell'attributo
+                if type(is_active_model) is not bool:
+                    is_active_model = False
+                    _logger.warning('ATTACHMENTS GC - The "_attachment_garbage_collector" attribute of the model "%s" is not boolean. The attriburte has been ignored.' % model_name)
+            
+            # Ignora come attivi i modelli nella safelist
+            if is_active_model and (model_name in cg_name_safelist):
                 is_active_model = False
-                _logger.warning('ATTACHMENTS GARBAGE COLLECTOR - The "_attachment_garbage_collector" attribute of the model "%s" is not boolean. The attriburte has been ignored.' % model_name)
-            # Ignora come attivi i modelli nella blacklist
-            if is_active_model and (model_name in cg_name_blacklist):
-                is_active_model = False
-                _logger.warning('ATTACHMENTS GARBAGE COLLECTOR - The model "%s" cannot have the "_attachment_garbage_collector" attribute set to True. The attriburte has been ignored.' % model_name)
+                _logger.warning('ATTACHMENTS GC - According to the safelist, the model "%s" cannot be active. The attachments of this model have been ignored.' % model_name)
 
             # Se punta a "ir.attachment" o è un modello attivo
             if m2m_attachment_fields or m2o_attachment_fields:
@@ -143,7 +100,15 @@ class IrAttachment(models.Model):
             if is_active_model:
                 act_models.append(model_name)
 
-        _logger.info('ATTACHMENTS GARBAGE COLLECTOR -  %d  Many2many and  %d  Many2one fields related to "ir.attachment" were found on  %d  different models.' % (count['m2m'], count['m2o'], count['model']))
+            if is_active_model and not (m2m_attachment_fields or m2o_attachment_fields):
+                _logger.warning('ATTACHMENTS GC - The active model "%s" does not appear to have any relational fields (m2m or m2o) pointing to "ir.attachment".' % model_name)
+
+        _logger.info('ATTACHMENTS GC - %d  Many2many and  %d  Many2one fields related to "ir.attachment" were found on  %d  different models.' % (count['m2m'], count['m2o'], count['model']))
+        _logger.info('ATTACHMENTS GC - The service %s on this  %d  models: %s' % (
+            ('is forced to be active' if force_models else 'is active'),
+            len(act_models),
+            ', '.join(act_models))
+        )
 
         # Tutti gli attachment associati ai modelli "attivi" con "res_model" ma senza "res_id"
         all_active_models_attachments = self.sudo().with_context(active_test=False).search([
@@ -160,8 +125,7 @@ class IrAttachment(models.Model):
                     #     mai accadere. Controllare di nuovo in futuro.
         ])
         
-        _logger.info('ATTACHMENTS GARBAGE COLLECTOR - The service is active this  %d  models: %s' % (len(act_models), ', '.join(act_models)))
-        _logger.info('ATTACHMENTS GARBAGE COLLECTOR - In the active models,  %d  attachments were found which need to be checked.' % len(all_active_models_attachments))
+        _logger.info('ATTACHMENTS GC - In the active models,  %d  attachments were found which need to be checked.' % len(all_active_models_attachments))
 
         # Tutti gli attachment dei modelli "attivi" usati da record di qualunque altro modello
         used_attachments = self.env['ir.attachment'] # empty recordset
@@ -197,18 +161,29 @@ class IrAttachment(models.Model):
         count['poa'] = len(maybe_orphan_attachments)
 
         # Esegue le operazioni
-        confirmed_orphan_attachments.sudo().unlink()
+        if do_unlink:
+            confirmed_orphan_attachments.sudo().unlink()
         maybe_orphan_attachments.sudo().write({'maybe_orphan': True})
         not_orphan_attachments.sudo().write({'maybe_orphan': False})
 
-        _logger.info('ATTACHMENTS GARBAGE COLLECTOR - DONE')
-        _logger.info('                              - Found  %d  possible orphan attachments on the active model/s.' % count['oa'])
-        _logger.info('                              -        %d  confirmed orphan attachments will be DELETED.' % count['coa'])
-        _logger.info('                              -        %d  new possible orphan attachments will be MARKED.' % count['poa'])
+        _logger.info('ATTACHMENTS GC - DONE.')
+        _logger.info('ATTACHMENTS GC - Found  %d  possible orphan attachments on the active models.' % count['oa'])
+        if do_unlink:
+            _logger.info('ATTACHMENTS GC - DELETED  %d  confirmed orphan attachments.' % count['coa'])
+        else:
+            _logger.info('ATTACHMENTS GC - SKIPPED  %d  confirmed orphan attachments (not deleted).' % count['coa'])
+        _logger.info('ATTACHMENTS GC - MARKED  %d  new possible orphan attachments.' % count['poa'])
 
-        # NB:
-        # Gli attachment rimasti orfani durante la creazione di un messaggio/nota/email
-        # (quando è ancora in un transient) che alla fine non è andata a buon fine,
-        # sono già eliminati da self.emv['mail.thread']._garbage_collect_attachments()
-        # In pratica elimina gli attachment con('res_model', '=', 'mail.compose.message')
-        # creati o modificati al minimo un giorno prima.
+        return True
+
+        # REMEMBER THAT...
+        # Attachments left orphaned during the creation of a message/note/email
+        # (while still in a transient) that eventually failed are already deleted by
+        # self.emv['mail.thread']._garbage_collect_attachments(). In practice, this
+        # deletes attachments with ('res_model', '=', 'mail.compose.message') created
+        # or modified at least one day before.
+
+
+    # @todo: Implementare registro/log degli attachment cancellati (hash, tipo, dimensione, res_name, res_model)
+    # @todo: Implementare action multi "set as not orphan"
+    # @todo: ?? Aggiungere constraint che dà errore se si marcano record con "res_id" != False ??
